Finished the gouraud RGB better method. A bit of copy of sourcePtrs from R channel to G to B before every draw call of the same scanline CELs. Can be optimized more.

+++ src/project/experiments/common/engine_mesh.h
@@ -100,7 +100,7 @@ void setAllPolyData(Mesh *ms, int numPoints, int textureId, int palId);
 void updateAllPolyTextureId(Mesh *ms, int textureId);
 void updatePolyTexData(Mesh *ms);
 void flipMeshPolyOrder(Mesh *ms);
-void scaleMesh(Mesh *ms, int scaleX, int scaleY, int scaleZ);
+void scaleMesh(Mesh* ms, float scaleX, float scaleY, float scaleZ);
 void flipMeshVerticesIfNeg(Mesh *ms, bool flipX, int flipY, bool flipZ);
 
 void updateMeshCELs(Mesh *ms);
